# docs/openapi.yaml
openapi: 3.0.3
info:
  title: TCGer API
  description: REST API for TCGer - Trading Card Collection Manager
  version: 0.1.0
servers:
  - url: /api
    description: Docker gateway / reverse-proxy path
  - url: /
    description: Direct backend service (no API prefix)
  - url: http://localhost:3004
    description: Local backend example port
security: []
...
paths:
...
  /cards/search:
    get:
      tags: [Cards]
      summary: Search cards across all TCGs
      operationId: searchCards
      security:
        - bearerAuth: []
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
            minLength: 1
        - name: tcg
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/TcgCode"
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  cards:
                    type: array
                    items:
                      $ref: "#/components/schemas/Card"
                  total:
                    type: integer
        "400":
          $ref: "#/components/responses/ValidationError"
  /cards/search/all:
    get:
      tags: [Cards]
      summary: Exhaustive name search — every printing matching a name
      description: >
        Unlike `/cards/search`, which returns a capped preview page, this pages
        through the provider for every match. Used to expand wishlist rules such
        as "every Darkrai". `limit` applies per game, so an all-games search can
        return up to `limit × games` cards.
      operationId: searchAllCards
      security:
        - bearerAuth: []
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
            minLength: 1
        - name: tcg
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/TcgCode"
        - name: unique
          in: query
          required: false
          description: >
            `prints` returns every printing of every matching card; `cards`
            collapses them to one entry per distinct card.
          schema:
            type: string
            enum: [prints, cards]
            default: prints
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 500
      responses:
        "200":
          description: Every matching card, up to the limit
          content:
            application/json:
              schema:
                type: object
                properties:
                  cards:
                    type: array
                    items:
                      $ref: "#/components/schemas/Card"
                  total:
                    type: integer
        "400":
          $ref: "#/components/responses/ValidationError"
  /cards/{tcg}/{cardId}:
    get:
      tags: [Cards]
      summary: Get a single card by TCG and ID
      operationId: getCard
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/tcg"
        - $ref: "#/components/parameters/cardId"
      responses:
        "200":
          description: Card detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  card:
                    $ref: "#/components/schemas/Card"
        "404":
          $ref: "#/components/responses/NotFound"
  /cards/{tcg}/{cardId}/prints:
    get:
      tags: [Cards]
      summary: Get all prints of a card
      operationId: getCardPrints
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/tcg"
        - $ref: "#/components/parameters/cardId"
      responses:
        "200":
          description: Card prints
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardPrintsResponse"

  # ─── Collections ─────────────────────────────────────────────────────
...
  /wishlists:
    get:
      tags: [Wishlists]
      summary: List all wishlists for the authenticated user
      operationId: getWishlists
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of wishlists with completion data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WishlistResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Wishlists]
      summary: Create a new wishlist
      operationId: createWishlist
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWishlistInput"
      responses:
        "201":
          description: Wishlist created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WishlistResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /wishlists/{wishlistId}:
    get:
      tags: [Wishlists]
      summary: Get a single wishlist with all cards and completion status
      operationId: getWishlist
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/wishlistId"
      responses:
        "200":
          description: Wishlist detail with cards and ownership info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WishlistResponse"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Wishlists]
      summary: Update a wishlist
      operationId: updateWishlist
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/wishlistId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWishlistInput"
      responses:
        "200":
          description: Updated wishlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WishlistResponse"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Wishlists]
      summary: Delete a wishlist
      operationId: deleteWishlist
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/wishlistId"
      responses:
        "204":
          description: Wishlist deleted
        "404":
          $ref: "#/components/responses/NotFound"
  /wishlists/{wishlistId}/cards:
    post:
      tags: [Wishlists]
      summary: Add a card to a wishlist
      operationId: addCardToWishlist
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/wishlistId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddWishlistCardInput"
      responses:
        "201":
          description: Card added to wishlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WishlistCardResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /wishlists/{wishlistId}/cards/{wishlistCardId}:
    delete:
      tags: [Wishlists]
      summary: Remove a card from a wishlist
      operationId: removeCardFromWishlist
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/wishlistId"
        - name: wishlistCardId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Card removed from wishlist
        "404":
          $ref: "#/components/responses/NotFound"
  /wishlists/{wishlistId}/rules:
    post:
      tags: [Wishlists]
      summary: Add (or refresh) an expansion rule on a wishlist
      description: >
        Rules describe the cards a wishlist wants in the abstract — "every
        Darkrai in Pokémon", "everything in Prismatic Evolutions". Posting a
        rule that already exists refreshes it instead of duplicating it.
        Expansion happens client-side: fetch matches from `/cards/search/all`
        or `/cards/sets/{tcg}/{setCode}`, then POST them to
        `/wishlists/{wishlistId}/cards/batch`.
      operationId: addWishlistRule
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/wishlistId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWishlistRuleInput"
      responses:
        "201":
          description: Rule saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WishlistRuleResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFound"
  /wishlists/{wishlistId}/rules/{ruleId}:
    patch:
      tags: [Wishlists]
      summary: Update a rule (used to record sync results)
      operationId: updateWishlistRule
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/wishlistId"
        - name: ruleId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWishlistRuleInput"
      responses:
        "200":
          description: Updated rule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WishlistRuleResponse"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Wishlists]
      summary: Remove a rule (cards it already added stay on the wishlist)
      operationId: removeWishlistRule
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/wishlistId"
        - name: ruleId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Rule removed
        "404":
          $ref: "#/components/responses/NotFound"

  # ─── Users ───────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    tcg:
      name: tcg
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/TcgCode"
    cardId:
      name: cardId
      in: path
      required: true
      schema:
        type: string
...
    wishlistId:
      name: wishlistId
      in: path
      required: true
      schema:
        type: string
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    TcgCode:
      type: string
      enum: [yugioh, magic, pokemon, onepiece, lorcana, dragonball]
...
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          example: NOT_FOUND
        message:
          type: string
          example: Resource not found
        details:
          type: object

    # ─── Auth ────────────────────────────────────────────────────────
...
    Card:
      type: object
      required: [id, tcg, name]
      properties:
        id:
          type: string
        tcg:
          $ref: "#/components/schemas/TcgCode"
        name:
          type: string
        setCode:
          type: string
        setName:
          type: string
        rarity:
          type: string
        collectorNumber:
          type: string
        releasedAt:
          type: string
        imageUrl:
          type: string
        imageUrlSmall:
          type: string
        setSymbolUrl:
          type: string
          description: URL to the set expansion symbol image
        setLogoUrl:
          type: string
          description: URL to the set logo/branding image (Pokemon and MTG)
        regulationMark:
          type: string
        language:
          type: string
        pokemonPrint:
          $ref: "#/components/schemas/PokemonPrintMetadata"
        attributes:
          type: object
          additionalProperties: true
    PokemonPrintMetadata:
      type: object
      properties:
        tcgdexId:
          type: string
        tcgdexImage:
          type: string
        variants:
          $ref: "#/components/schemas/PokemonVariantFlags"
        finishes:
          type: array
          items:
            type: string
            enum: [normal, reverse, holo, firstEdition]
        category:
          type: string
        regulationMark:
          type: string
        language:
          type: string
    PokemonVariantFlags:
      type: object
      properties:
        normal:
          type: boolean
        reverse:
          type: boolean
        holo:
          type: boolean
        firstEdition:
          type: boolean
    PokemonFunctionalGroup:
      type: object
      required: [functionalKey, name]
      properties:
        functionalKey:
          type: string
        name:
          type: string
        supertype:
          type: string
        subtypes:
          type: array
          items:
            type: string
        hp:
          type: string
        regulationMark:
          type: string
        category:
          type: string
        normalizedRules:
          type: string
          nullable: true
        attacks:
          type: array
          items:
            type: object
            required: [name]
            properties:
              name:
                type: string
              cost:
                type: array
                items:
                  type: string
              text:
                type: string
                nullable: true
              damage:
                type: string
                nullable: true
              convertedEnergyCost:
                type: integer
        abilities:
          type: array
          items:
            type: object
            required: [name]
            properties:
              name:
                type: string
              text:
                type: string
                nullable: true
              type:
                type: string
        rules:
          type: array
          nullable: true
          items:
            type: string
    CardPrintsResponse:
      oneOf:
        - type: object
          required: [mode, prints, total]
          properties:
            mode:
              type: string
              enum: [simple]
            prints:
              type: array
              items:
                $ref: "#/components/schemas/Card"
            total:
              type: integer
        - type: object
          required: [mode, prints, total, functionalGroup]
          properties:
            mode:
              type: string
              enum: [pokemon-functional]
            prints:
              type: array
              items:
                $ref: "#/components/schemas/Card"
            total:
              type: integer
            functionalGroup:
              $ref: "#/components/schemas/PokemonFunctionalGroup"
      discriminator:
        propertyName: mode

    # ─── Collections ─────────────────────────────────────────────────
...
    WishlistCardResponse:
      type: object
      required: [id, externalId, tcg, name, owned, ownedQuantity, createdAt]
      properties:
        id:
          type: string
        externalId:
          type: string
        tcg:
          $ref: "#/components/schemas/TcgCode"
        name:
          type: string
        setCode:
          type: string
        setName:
          type: string
        rarity:
          type: string
        imageUrl:
          type: string
        imageUrlSmall:
          type: string
        setSymbolUrl:
          type: string
          description: URL to the set expansion symbol image
        setLogoUrl:
          type: string
          description: URL to the set logo/branding image
        collectorNumber:
          type: string
        notes:
          type: string
        owned:
          type: boolean
          description: Whether this card exists in the user's collection
        ownedQuantity:
          type: integer
          description: Total quantity owned across all binders
        createdAt:
          type: string
          format: date-time
    WishlistResponse:
      type: object
      required: [id, name, cards, rules, totalCards, ownedCards, completionPercent, createdAt, updatedAt]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        colorHex:
          type: string
        cards:
          type: array
          items:
            $ref: "#/components/schemas/WishlistCardResponse"
        rules:
          type: array
          description: Saved expansion rules; empty for a purely manual wishlist
          items:
            $ref: "#/components/schemas/WishlistRuleResponse"
        totalCards:
          type: integer
          description: Number of unique cards in the wishlist
        ownedCards:
          type: integer
          description: Number of cards that exist in the user's collection
        completionPercent:
          type: integer
          description: Completion percentage (0-100)
          minimum: 0
          maximum: 100
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    WishlistRuleResponse:
      type: object
      required: [id, type, includeAllPrintings, autoSync, createdAt, updatedAt]
      properties:
        id:
          type: string
        type:
          type: string
          enum: [name, set]
        tcg:
          $ref: "#/components/schemas/TcgCode"
        query:
          type: string
          description: Name fragment matched by a name rule
        setCode:
          type: string
        setName:
          type: string
        includeAllPrintings:
          type: boolean
          description: Match every printing rather than one entry per distinct card
        autoSync:
          type: boolean
        lastSyncedAt:
          type: string
          format: date-time
        lastMatchCount:
          type: integer
          description: How many cards the rule matched on its last sync
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateWishlistRuleInput:
      type: object
      required: [type]
      properties:
        type:
          type: string
          enum: [name, set]
        tcg:
          $ref: "#/components/schemas/TcgCode"
        query:
          type: string
          description: Required for a name rule
          minLength: 1
        setCode:
          type: string
          description: Required for a set rule (along with tcg)
          minLength: 1
        setName:
          type: string
        includeAllPrintings:
          type: boolean
          default: true
        autoSync:
          type: boolean
          default: true
    UpdateWishlistRuleInput:
      type: object
      properties:
        autoSync:
          type: boolean
        includeAllPrintings:
          type: boolean
        lastSyncedAt:
          type: string
          format: date-time
        lastMatchCount:
          type: integer
          minimum: 0
    CreateWishlistInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        colorHex:
          type: string
          pattern: "^([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$"
    UpdateWishlistInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        colorHex:
          type: string
          pattern: "^([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$"
    AddWishlistCardInput:
      type: object
      required: [externalId, tcg, name]
      properties:
        externalId:
          type: string
          minLength: 1
        tcg:
          type: string
          minLength: 1
        name:
          type: string
          minLength: 1
        setCode:
          type: string
        setName:
          type: string
        rarity:
          type: string
        imageUrl:
          type: string
        imageUrlSmall:
          type: string
        setSymbolUrl:
          type: string
        setLogoUrl:
          type: string
        collectorNumber:
          type: string
        notes:
          type: string

    # ─── Users ───────────────────────────────────────────────────────
...